refactor: extract month grid UI
